fix: create archive camera.hpp and define shapes
